feat: change bnb_4bit_compute_dtype from f16 to f32

+++ colab_bridge/standalone_llm_bridge.py
@@ -283,7 +283,7 @@ class LLMEngine:
                     load_in_4bit=True,
                     bnb_4bit_use_double_quant=True,
                     bnb_4bit_quant_type="nf4",
-                    bnb_4bit_compute_dtype=torch.float16,
+                    bnb_4bit_compute_dtype=torch.float32,
                 )
             else:
                 model_kwargs["device_map"] = "cpu"
